isr done, wip irq

+++ kernel/include/kernel/framebuffer.h
@@ -6,7 +6,7 @@ typedef struct framebufferinfo_t {
     unsigned int hres; // The horizontal resolution of the framebuffer
     unsigned int vres; // The vertical resolution of the framebuffer
     unsigned int pitch; // The pixels per scanline in the framebuffer
-    unsigned long long size; // The size of the framebuffer
+    unsigned long int size; // The size of the framebuffer
 } framebufferinfo_t;
 
 // Initialize the framebuffer
